Separation of datasource and widget protocols
